Bring getPrivateChannel and getGroup up also.
