+++ src/main/java/net/dv8tion/jda/core/events/message/GenericMessageEvent.java
@@ -15,10 +15,12 @@
  */
 package net.dv8tion.jda.core.events.message;
 
+import net.dv8tion.jda.client.entities.Group;
 import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.entities.ChannelType;
 import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.MessageChannel;
+import net.dv8tion.jda.core.entities.PrivateChannel;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.events.Event;
 
@@ -78,5 +80,47 @@ public abstract class GenericMessageEvent extends Event
     public Guild getGuild()
     {
         return isFromType(ChannelType.TEXT) ? ((TextChannel) channel).getGuild() : null;
+    }
+
+    /**
+     * Returns the {@link net.dv8tion.jda.core.entities.TextChannel TextChannel} the Message was received in.
+     * <br>If this Message was not received in a {@link net.dv8tion.jda.core.entities.TextChannel TextChannel},
+     * this will return {@code null}.
+     *
+     * @return The TextChannel the Message was received in or null if not from a TextChannel
+     *
+     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
+     */
+    public TextChannel getTextChannel()
+    {
+        return isFromType(ChannelType.TEXT) ? (TextChannel) channel : null;
+    }
+
+    /**
+     * Returns the {@link net.dv8tion.jda.core.entities.PrivateChannel PrivateChannel} the Message was received in.
+     * <br>If this Message was not received in a {@link net.dv8tion.jda.core.entities.PrivateChannel PrivateChannel},
+     * this will return {@code null}.
+     *
+     * @return The PrivateChannel the Message was received in or null if not from a PrivateChannel
+     *
+     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
+     */
+    public PrivateChannel getPrivateChannel()
+    {
+        return isFromType(ChannelType.PRIVATE) ? (PrivateChannel) channel : null;
+    }
+
+    /**
+     * Returns the {@link net.dv8tion.jda.client.entities.Group Group} the Message was received in.
+     * <br>If this Message was not received in a {@link net.dv8tion.jda.client.entities.Group Group},
+     * this will return {@code null}.
+     *
+     * @return The Group the Message was received in or null if not from a Group
+     *
+     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
+     */
+    public Group getGroup()
+    {
+        return isFromType(ChannelType.GROUP) ? (Group) channel : null;
     }
 }
