+++ src/main/java/net/dv8tion/jda/core/events/message/MessageReceivedEvent.java
@@ -15,7 +15,6 @@
  */
 package net.dv8tion.jda.core.events.message;
 
-import net.dv8tion.jda.client.entities.Group;
 import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.entities.*;
 
@@ -75,48 +74,6 @@ public class MessageReceivedEvent extends GenericMessageEvent
     public Member getMember()
     {
         return isFromType(ChannelType.TEXT) && !isWebhookMessage() ? getGuild().getMember(getAuthor()) : null;
-    }
-
-    /**
-     * Returns the {@link net.dv8tion.jda.core.entities.PrivateChannel PrivateChannel} the Message was received in.
-     * <br>If this Message was not received in a {@link net.dv8tion.jda.core.entities.PrivateChannel PrivateChannel},
-     * this will return {@code null}.
-     *
-     * @return The PrivateChannel the Message was received in or null if not from a PrivateChannel
-     *
-     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
-     */
-    public PrivateChannel getPrivateChannel()
-    {
-        return message.getPrivateChannel();
-    }
-
-    /**
-     * Returns the {@link net.dv8tion.jda.client.entities.Group Group} the Message was received in.
-     * <br>If this Message was not received in a {@link net.dv8tion.jda.client.entities.Group Group},
-     * this will return {@code null}.
-     *
-     * @return The Group the Message was received in or null if not from a Group
-     *
-     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
-     */
-    public Group getGroup()
-    {
-        return message.getGroup();
-    }
-
-    /**
-     * Returns the {@link net.dv8tion.jda.core.entities.TextChannel TextChannel} the Message was received in.
-     * <br>If this Message was not received in a {@link net.dv8tion.jda.core.entities.TextChannel TextChannel},
-     * this will return {@code null}.
-     *
-     * @return The TextChannel the Message was received in or null if not from a TextChannel
-     *
-     * @see net.dv8tion.jda.core.events.message.GenericMessageEvent#isFromType(ChannelType)
-     */
-    public TextChannel getTextChannel()
-    {
-        return message.getTextChannel();
     }
 
     /**
