+++ src/main/java/net/dv8tion/jda/core/events/message/MessageUpdateEvent.java
@@ -15,7 +15,6 @@
  */
 package net.dv8tion.jda.core.events.message;
 
-import net.dv8tion.jda.client.entities.Group;
 import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.entities.*;
 
@@ -51,20 +50,5 @@ public class MessageUpdateEvent extends GenericMessageEvent
     public Member getMember()
     {
         return  isFromType(ChannelType.TEXT) ? getGuild().getMember(getAuthor()) : null;
-    }
-
-    public PrivateChannel getPrivateChannel()
-    {
-        return message.getPrivateChannel();
-    }
-
-    public Group getGroup()
-    {
-        return message.getGroup();
-    }
-
-    public TextChannel getTextChannel()
-    {
-        return message.getTextChannel();
     }
 }
